+++ projects/googleplacesautocomplete/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/googleplacesautocomplete",
+  "dest": "../../dist/ngx-googleplacesautocomplete",
   "lib": {
     "entryFile": "src/public-api.ts"
   }
