+++ projects/googleplacesautocomplete/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "googleplacesautocomplete",
+  "name": "@asif/ngx-googleplacesautocomplete",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^14.2.0",
