--- projects/googleplacesautocomplete/src/lib/google-places-autocomplete/interfaces/IOptions.ts
@@ -1,4 +1,0 @@
-export interface IOptions{
-    fields:Array<string> | ['ALL'];
-    types:Array<any>
-}
