+++ projects/googleplacesautocomplete/src/lib/ngx-google-places-autocomplete/interfaces/IOptions.ts
@@ -1,0 +1,10 @@
+export interface IOptions{
+    fields:Array<string> | ['ALL'];
+    types:Array<any>
+}
+export interface IBounds{
+    east:number
+    west:number
+    north:number
+    south:number
+}
