+++ projects/googleplacesautocomplete/src/lib/ngx-google-places-autocomplete/interfaces/IPlace.ts
