+++ projects/googleplacesautocomplete/src/public-api.ts
@@ -2,4 +2,6 @@
  * Public API Surface of googleplacesautocomplete
  */
 
-export * from './lib/google-places-autocomplete/google-places-autocomplete.module';
+export * from './lib/ngx-google-places-autocomplete/ngx-google-places-autocomplete.module';
+export * from './lib/ngx-google-places-autocomplete/ngx-autocomplete.directive';
+export * from './lib/ngx-google-places-autocomplete/ngx-google-places-autocomplete.service';
